Add sbt-scource-format sbt plugin
This is a simple sbt plugin that adds javafmt and clangfmt tasks that
are analogous to the scalafmt task from the namesake plugin. For each of
the tasks, the source files can be configured with the
(clangfmt|javafmt)Source keys. I added some simple scripted tests that
both format and check modes work. This commit will also server as the
v0.1.0 release which I will publish to maven.